requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.5.2
    # via
    #   django
    #   djoser
asttokens==2.0.5
    # via stack-data
attrs==22.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
black==22.6.0
    # via -r requirements.in
bleach==5.0.1
    # via nbconvert
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via black
colorama==0.4.5
    # via
    #   click
    #   ipython
commonmark==0.9.1
    # via rich
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==37.0.4
    # via social-auth-core
debugpy==1.6.2
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via
    #   nbconvert
    #   python3-openid
    #   social-auth-core
django==4.0.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-extensions==3.2.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via
    #   -r requirements.in
    #   djoser
djoser==2.1.0
    # via -r requirements.in
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.9.1
    # via stack-data
fastjsonschema==2.16.1
    # via nbformat
flake8==4.0.1
    # via -r requirements.in
idna==3.3
    # via requests
ipykernel==6.15.1
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.4.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
    #   qtconsole
ipywidgets==7.7.1
    # via jupyter
itypes==1.2.0
    # via coreapi
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   coreschema
    #   nbconvert
    #   notebook
jsonschema==4.9.1
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.3.4
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.4
    # via jupyter
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==1.1.1
    # via ipywidgets
markdown==3.4.1
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mccabe==0.6.1
    # via flake8
mistune==0.8.4
    # via nbconvert
mypy-extensions==0.4.3
    # via black
nbclient==0.6.6
    # via nbconvert
nbconvert==6.5.0
    # via
    #   jupyter
    #   notebook
nbformat==5.4.0
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via
    #   jupyter
    #   widgetsnbextension
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via
    #   ipykernel
    #   nbconvert
    #   qtpy
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.2
    # via black
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.30
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.1
    # via ipykernel
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==2.4.0
    # via flake8
pygments==2.12.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
    #   rich
pyjwt==2.4.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via jupyter-client
python-dotenv==0.20.0
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.1
    # via djangorestframework
pywin32==304
    # via jupyter-core
pywinpty==2.0.6
    # via terminado
pyzmq==23.2.0
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.1
    # via jupyter
qtpy==2.1.0
    # via qtconsole
requests==2.28.1
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rich==12.5.1
    # via -r requirements.in
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
stack-data==0.3.0
    # via ipython
terminado==0.15.0
    # via notebook
tinycss2==1.1.1
    # via nbconvert
tomli==2.0.1
    # via black
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.3.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
tzdata==2022.1
    # via django
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.11
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==3.6.1
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
